requirements.txt
# Core Data & Modeling
pandas==2.2.3
numpy==2.2.3
scikit-learn==1.6.1
scipy==1.15.2
python-dateutil==2.9.0.post0
pytz==2025.1
joblib==1.4.2
threadpoolctl==3.5.0

# Visualization
matplotlib==3.10.0
seaborn==0.13.2
plotly==6.0.0
pydeck==0.9.1
altair==5.5.0

# Streamlit App
streamlit==1.44.1

# File Handling
openpyxl==3.1.5
pyarrow==19.0.1
et_xmlfile==2.0.0
PyYAML==6.0.2

# Web + API
requests==2.32.3
urllib3==2.3.0
idna==3.10
charset-normalizer==3.4.1

# Google Cloud Platform (✅ GCP-specific additions)
google-cloud-storage==2.16.0
google-cloud-bigquery==3.17.1

# Snowflake Connectors (✅ retained for backward compatibility)
snowflake-connector-python==3.13.2
snowflake-sqlalchemy==1.7.3

# Dependencies & Utilities
typing_extensions==4.12.2
packaging==24.2
six==1.17.0
MarkupSafe==3.0.2
click==8.1.8
fonttools==4.56.0
kiwisolver==1.4.8
cycler==0.12.1
attrs==25.3.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
referencing==0.36.2
rpds-py==0.24.0
iniconfig==2.1.0
pluggy==1.5.0
toml==0.10.2
tomlkit==0.13.2
platformdirs==4.3.6
filelock==3.17.0
cachetools==5.5.2
protobuf==5.29.4
blinker==1.9.0
tenacity==9.1.2

# Crypto-related (used by Snowflake/PyJWT)
asn1crypto==1.5.1
cffi==1.17.1
pycparser==2.22
cryptography==44.0.1
pyOpenSSL==24.3.0
PyJWT==2.10.1

# Git Integration
gitdb==4.0.12
GitPython==3.1.44
smmap==5.0.2

# Misc
Calculator==1.1.0
narwhals==1.27.1python-dotenv==1.1.1
